# kvclient: add `ReturnOnRangeBoundary` request field 