This patch adds an option `ReturnOnRangeBoundary` which causes
DistSender to return partial non-empty responses on range boundaries
with reason `RESUME_RANGE_BOUNDARY`. This avoids sending requests with
small remaining limits to subsequent ranges, which would increase the
overall number of range RPC requests. It is opt-in because it could be
detrimental in some cases, e.g. with large batches of `Get` requests.

A version gate is not included, since this change is entirely on the KV
client side: the caller and the DistSender are assumed to be upgraded in
sync.

Resolves #70564.
Touches #47044.

Release note: None