@erikgrinaker 2021-10-11
Wrote a [benchmark](https://github.com/cockroachdb/cockroach/blob/1bd6317f4c03dc7c2145c4f19e28d2c94c33ccd2/pkg/kv/kvclient/kvcoord/dist_sender_server_test.go#L3157-L3158) to compare the cost of additional DistSender RPCs against the cost of additional sender stack traversals when scanning across a set of ranges within a transaction. Parameters:

```go
Ranges                = 10   // number of ranges to create
KeysPerRange          = 9    // number of keys to write per range
MaxSpanRequestKeys    = 10   // max number of keys in each scan response
ReturnOnRangeBoundary = true // if true, enable ReturnOnRangeBoundary
Latency               = 0    // additional RPC request latency
```

This favors `ReturnOnRangeBoundary`, since the old behavior will have to do about twice as many RPC requests overall -- it will e.g. fetch 9 rows from range 1 and then go to range 2 only to fetch a single row. Results, where old/new is `ReturnOnRangeBoundary = false/true`, and the `Latency` results are with additional RPC latency (1ms is roughly inter-AZ latency, 5ms simulates additional IO latency):

```
name                              old time/op  new time/op  delta
ReturnOnRangeBoundary              616µs ± 4%   386µs ± 3%  -37.31%  (p=0.000 n=9+10)
ReturnOnRangeBoundary-1msLatency  23.7ms ± 3%  13.5ms ± 3%  -42.99%  (p=0.000 n=10+9)
ReturnOnRangeBoundary-5msLatency   102ms ± 2%    57ms ± 1%  -44.45%  (p=0.000 n=10+9)
```

If we instead favor the old behavior by allowing it to gather all results in one go with `MaxSpanRequestKeys = 100` while the new behavior has to traverse the sender stack for each range, we find:

```
name                              old time/op  new time/op  delta
ReturnOnRangeBoundary              311µs ± 3%   386µs ± 1%  +24.19%  (p=0.000 n=10+8)
ReturnOnRangeBoundary-1msLatency  13.0ms ± 1%  13.8ms ± 8%  +6.71%   (p=0.000 n=9+10)
ReturnOnRangeBoundary-5msLatency  55.6ms ± 2%  56.0ms ± 1%   ~       (p=0.113 n=10+9)
```

Thus, with negligible request latency the tradeoff is highly dependant on how the spans, limits and range boundaries overlap, which the caller must take into account. However, as request latency increases, the relative worst-case cost of `ReturnOnRangeBoundary` approaches 0.