@erikgrinaker 2021-11-01
Friendly ping for reviews.

@andreimatei Can you look this over, or let me know if I should reassign to someone else on KV?
@dt Does this cover Bulk's needs?
@RaduBerinde Does this cover SQL's needs for #47044?