@dt 2021-11-09
> @dt Does this cover Bulk's needs?
Looks great, should get us the current behavior but without tricking with ExportResponse's fake sizes -- so much cleaner, thanks!

(I only looked at the diff in api.proto and dist_sender.go, so leaving review/approving to a KV reviewer)